requirements.txt
openai==1.14.3
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
pinecone-client
transformers
pytorch
scikit-learn